# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow Pro - Project Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ProjectPriority =
  (typeof ProjectPriority)[keyof typeof ProjectPriority];

export const ProjectPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  priority: ProjectPriority;
  progress: number;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type CreateProjectBodyPriority =
  (typeof CreateProjectBodyPriority)[keyof typeof CreateProjectBodyPriority];

export const CreateProjectBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
  status: CreateProjectBodyStatus;
  priority: CreateProjectBodyPriority;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  planning: "planning",
  active: "active",
  on_hold: "on_hold",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type UpdateProjectBodyPriority =
  (typeof UpdateProjectBodyPriority)[keyof typeof UpdateProjectBodyPriority];

export const UpdateProjectBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateProjectBodyStatus;
  priority?: UpdateProjectBodyPriority;
  progress?: number;
  /** @nullable */
  dueDate?: string | null;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  review: "review",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  review: "review",
  done: "done",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status: CreateTaskBodyStatus;
  priority: CreateTaskBodyPriority;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  review: "review",
  done: "done",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type TeamMemberRole =
  (typeof TeamMemberRole)[keyof typeof TeamMemberRole];

export const TeamMemberRole = {
  admin: "admin",
  manager: "manager",
  developer: "developer",
  designer: "designer",
  analyst: "analyst",
} as const;

export interface TeamMember {
  id: number;
  name: string;
  email: string;
  role: TeamMemberRole;
  /** @nullable */
  avatarUrl?: string | null;
  taskCount: number;
  createdAt: string;
}

export type CreateTeamMemberBodyRole =
  (typeof CreateTeamMemberBodyRole)[keyof typeof CreateTeamMemberBodyRole];

export const CreateTeamMemberBodyRole = {
  admin: "admin",
  manager: "manager",
  developer: "developer",
  designer: "designer",
  analyst: "analyst",
} as const;

export interface CreateTeamMemberBody {
  name: string;
  email: string;
  role: CreateTeamMemberBodyRole;
  /** @nullable */
  avatarUrl?: string | null;
}

export type UpdateTeamMemberBodyRole =
  (typeof UpdateTeamMemberBodyRole)[keyof typeof UpdateTeamMemberBodyRole];

export const UpdateTeamMemberBodyRole = {
  admin: "admin",
  manager: "manager",
  developer: "developer",
  designer: "designer",
  analyst: "analyst",
} as const;

export interface UpdateTeamMemberBody {
  name?: string;
  email?: string;
  role?: UpdateTeamMemberBodyRole;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface DashboardStats {
  totalProjects: number;
  activeProjects: number;
  completedProjects: number;
  totalTasks: number;
  completedTasks: number;
  inProgressTasks: number;
  overdueTasks: number;
  teamSize: number;
  completionRate: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  task_created: "task_created",
  task_completed: "task_completed",
  project_created: "project_created",
  project_updated: "project_updated",
  member_added: "member_added",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  entityName: string;
  actorName: string;
  createdAt: string;
}

export interface ProjectBreakdown {
  status: string;
  count: number;
  percentage: number;
}

export type ListProjectsParams = {
  status?: string;
  search?: string;
};

export type ListTasksParams = {
  projectId?: number;
  status?: string;
  priority?: string;
  assigneeId?: number;
};