# lib/api-zod/src/generated/types/createTeamMemberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow Pro - Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTeamMemberBodyRole } from "./createTeamMemberBodyRole";

export interface CreateTeamMemberBody {
  name: string;
  email: string;
  role: CreateTeamMemberBodyRole;
  /** @nullable */
  avatarUrl?: string | null;
}