# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow Pro - Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMemberRole } from "./teamMemberRole";

export interface TeamMember {
  id: number;
  name: string;
  email: string;
  role: TeamMemberRole;
  /** @nullable */
  avatarUrl?: string | null;
  taskCount: number;
  createdAt: string;
}