# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow Pro - Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyPriority } from "./updateProjectBodyPriority";
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateProjectBodyStatus;
  priority?: UpdateProjectBodyPriority;
  progress?: number;
  /** @nullable */
  dueDate?: string | null;
}