# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow Pro - Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}